+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added the `SQLiteQuery` task
 - Added style guidelines to the `README`
+- Source:
+  - `CloudForCustomers`
+- Added tasks `CloudForCustomersToDF`, `CloudForCustomersToCSV`
+- Added support for parquet in `CloudForCustomersToDF`
+- Added flow `CloudForCustomersToADLS`
 
 ### Changed
 - Modified `ADLSToAzureSQL` - *read_sep* and *write_sep* parameters added to the flow.